Core & Internals: Move REST utils; Fix #4035
to avoid accidential imports of server-side packages.
