+++ lib/rucio/web/rest/flaskapi/v1/temporary_did.py
@@ -23,8 +23,6 @@
 # - Muhammad Aditya Hilmy <didithilmy@gmail.com>, 2020
 # - Eli Chadwick <eli.chadwick@stfc.ac.uk>, 2020
 # - Benedikt Ziemons <benedikt.ziemons@cern.ch>, 2020
-#
-# PY3K COMPATIBLE
 
 from __future__ import print_function
 
@@ -36,8 +34,8 @@ from flask.views import MethodView
 
 from rucio.api.temporary_did import (add_temporary_dids)
 from rucio.common.exception import RucioException
-from rucio.common.utils import generate_http_error_flask
 from rucio.web.rest.flaskapi.v1.common import before_request, after_request
+from rucio.web.rest.utils import generate_http_error_flask
 
 
 class BulkDIDS(MethodView):
