+++ lib/rucio/web/rest/webpy/v1/importer.py
@@ -1,5 +1,6 @@
 #!/usr/bin/env python
-# Copyright 2012-2018 CERN for the benefit of the ATLAS collaboration.
+# -*- coding: utf-8 -*-
+# Copyright 2018-2020 CERN
 #
 # Licensed under the Apache License, Version 2.0 (the "License");
 # you may not use this file except in compliance with the License.
@@ -17,15 +18,15 @@
 # - Hannes Hansen <hannes.jakob.hansen@cern.ch>, 2018
 # - Andrew Lister <andrew.lister@stfc.ac.uk>, 2019
 # - Thomas Beermann <thomas.beermann@cern.ch>, 2020
-#
-# PY3K COMPATIBLE
+# - Benedikt Ziemons <benedikt.ziemons@cern.ch>, 2020
 
 from web import (application, data, header, loadhook, ctx, Created)
 
-from rucio.common.exception import RucioException
-from rucio.common.utils import generate_http_error, parse_response
 from rucio.api.importer import import_data
+from rucio.common.exception import RucioException
+from rucio.common.utils import parse_response
 from rucio.web.rest.common import rucio_loadhook, RucioController
+from rucio.web.rest.utils import generate_http_error
 
 URLS = (
     '/', 'Import',
