+++ lib/rucio/web/rest/webpy/v1/lifetime_exception.py
@@ -1,5 +1,6 @@
 #!/usr/bin/env python
-# Copyright 2012-2018 CERN for the benefit of the ATLAS collaboration.
+# -*- coding: utf-8 -*-
+# Copyright 2016-2020 CERN
 #
 # Licensed under the Apache License, Version 2.0 (the "License");
 # you may not use this file except in compliance with the License.
@@ -15,12 +16,11 @@
 #
 # Authors:
 # - Cedric Serfon <cedric.serfon@cern.ch>, 2016-2017
+# - Thomas Beermann <thomas.beermann@cern.ch>, 2018-2020
 # - Mario Lassnig <mario.lassnig@cern.ch>, 2018
-# - Hannes Hansen <hannes.jakob.hansen@cern.ch>, 2019
+# - Hannes Hansen <hannes.jakob.hansen@cern.ch>, 2018-2019
 # - Andrew Lister <andrew.lister@stfc.ac.uk>, 2019
-# - Thomas Beermann <thomas.beermann@cern.ch>, 2020
-#
-# PY3K COMPATIBLE
+# - Benedikt Ziemons <benedikt.ziemons@cern.ch>, 2020
 
 from json import loads, dumps
 
@@ -28,9 +28,9 @@ from web import application, ctx, data, header, BadRequest, Created, InternalErr
 
 from rucio.api.lifetime_exception import list_exceptions, add_exception, update_exception
 from rucio.common.exception import LifetimeExceptionNotFound, UnsupportedOperation, InvalidObject, RucioException, AccessDenied, LifetimeExceptionDuplicate
-from rucio.common.utils import generate_http_error, APIEncoder
+from rucio.common.utils import APIEncoder
 from rucio.web.rest.common import rucio_loadhook, check_accept_header_wrapper
-
+from rucio.web.rest.utils import generate_http_error
 
 URLS = ('/', 'LifetimeException',
         '/(.+)', 'LifetimeExceptionId',)
