+++ lib/rucio/web/rest/webpy/v1/temporary_did.py
@@ -1,5 +1,6 @@
 #!/usr/bin/env python
-# Copyright 2012-2018 CERN for the benefit of the ATLAS collaboration.
+# -*- coding: utf-8 -*-
+# Copyright 2016-2020 CERN
 #
 # Licensed under the Apache License, Version 2.0 (the "License");
 # you may not use this file except in compliance with the License.
@@ -15,23 +16,23 @@
 #
 # Authors:
 # - Vincent Garonne <vincent.garonne@cern.ch>, 2016
+# - Thomas Beermann <thomas.beermann@cern.ch>, 2018-2020
 # - Mario Lassnig <mario.lassnig@cern.ch>, 2018
 # - Hannes Hansen <hannes.jakob.hansen@cern.ch>, 2018
 # - Andrew Lister <andrew.lister@stfc.ac.uk>, 2019
-# - Thomas Beermann <thomas.beermann@cern.ch>, 2020
-#
-# PY3K COMPATIBLE
+# - Benedikt Ziemons <benedikt.ziemons@cern.ch>, 2020
 
 from __future__ import print_function
+
 from json import loads
 from traceback import format_exc
+
 from web import application, ctx, data, Created, InternalError, loadhook
 
 from rucio.api.temporary_did import (add_temporary_dids)
 from rucio.common.exception import RucioException
-from rucio.common.utils import generate_http_error
-
 from rucio.web.rest.common import rucio_loadhook, RucioController
+from rucio.web.rest.utils import generate_http_error
 
 URLS = ('', 'BulkDIDS',)
